requirements.txt
streamlit>=1.28.0
replicate>=0.15.0
requests>=2.31.0
Pillow>=10.0.0
pandas>=2.0.0
numpy>=1.24.0
faiss-cpu>=1.7.4
openai>=1.0.0
cloudinary>=1.36.0
python-dotenv>=1.0.0